added APP.js, move part of the code from index.js to App.js
